requirements.txt
anthropic>=0.25.0
spacy==3.7.4
fastapi>=0.110.0
uvicorn>=0.29.0
gradio>=4.0.0
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.7.0
numpy==1.26.4
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
